+++ clients/drcachesim/tracer/instru_online.cpp
@@ -150,7 +150,7 @@ online_instru_t::append_thread_header(byte *buf_ptr, thread_id_t tid)
     byte *new_buf = buf_ptr;
     new_buf += append_tid(new_buf, tid);
     new_buf += append_pid(new_buf, dr_get_process_id());
-    return (new_buf - buf_ptr);
+    return (int)(new_buf - buf_ptr);
 }
 
 int
@@ -162,7 +162,7 @@ online_instru_t::append_unit_header(byte *buf_ptr, thread_id_t tid)
                              // Truncated to 32 bits for 32-bit: we live with it.
                              (uintptr_t)instru_t::get_timestamp());
     new_buf += append_marker(new_buf, TRACE_MARKER_TYPE_CPU_ID, instru_t::get_cpu_id());
-    return (new_buf - buf_ptr);
+    return (int)(new_buf - buf_ptr);
 }
 
 void
